retry plugins: rename existing plugins to match extension name (#4887)
