+++ api/envoy/config/retry/previous_priorities/BUILD
@@ -3,8 +3,8 @@ licenses(["notice"])  # Apache 2
 load("//bazel:api_build_system.bzl", "api_proto_library_internal")
 
 api_proto_library_internal(
-    name = "other_priority",
-    srcs = ["other_priority_config.proto"],
+    name = "previous_priorities",
+    srcs = ["previous_priorities_config.proto"],
     deps = [
         "//envoy/api/v2/core:base",
     ],
