+++ source/extensions/retry/host/previous_hosts/BUILD
@@ -9,12 +9,10 @@ load(
 envoy_package()
 
 envoy_cc_library(
-    name = "other_priority_lib",
-    srcs = ["other_priority.cc"],
-    hdrs = ["other_priority.h"],
+    name = "previous_hosts_predicate_lib",
+    hdrs = ["previous_hosts.h"],
     deps = [
         "//include/envoy/upstream:retry_interface",
-        "//source/common/upstream:load_balancer_lib",
     ],
 )
 
@@ -23,11 +21,9 @@ envoy_cc_library(
     srcs = ["config.cc"],
     hdrs = ["config.h"],
     deps = [
-        ":other_priority_lib",
+        ":previous_hosts_predicate_lib",
         "//include/envoy/registry",
         "//include/envoy/upstream:retry_interface",
-        "//source/common/protobuf",
-        "//source/extensions/retry/priority:well_known_names",
-        "@envoy_api//envoy/config/retry/other_priority:other_priority_cc",
+        "//source/extensions/retry/host:well_known_names",
     ],
 )
