+++ source/extensions/retry/host/previous_hosts/config.h
@@ -2,7 +2,7 @@
 
 #include "envoy/upstream/retry.h"
 
-#include "extensions/retry/host/other_hosts/other_hosts.h"
+#include "extensions/retry/host/previous_hosts/previous_hosts.h"
 #include "extensions/retry/host/well_known_names.h"
 
 namespace Envoy {
@@ -10,11 +10,11 @@ namespace Extensions {
 namespace Retry {
 namespace Host {
 
-class OtherHostsRetryPredicateFactory : public Upstream::RetryHostPredicateFactory {
+class PreviousHostsRetryPredicateFactory : public Upstream::RetryHostPredicateFactory {
 public:
   Upstream::RetryHostPredicateSharedPtr createHostPredicate(const Protobuf::Message&,
                                                             uint32_t retry_count) override {
-    return std::make_shared<OtherHostsRetryPredicate>(retry_count);
+    return std::make_shared<PreviousHostsRetryPredicate>(retry_count);
   }
 
   std::string name() override { return RetryHostPredicateValues::get().PreviousHostsPredicate; }
