+++ source/extensions/retry/host/previous_hosts/previous_hosts.h
@@ -4,9 +4,9 @@
 #include "envoy/upstream/upstream.h"
 
 namespace Envoy {
-class OtherHostsRetryPredicate : public Upstream::RetryHostPredicate {
+class PreviousHostsRetryPredicate : public Upstream::RetryHostPredicate {
 public:
-  OtherHostsRetryPredicate(uint32_t retry_count) : attempted_hosts_(retry_count) {}
+  PreviousHostsRetryPredicate(uint32_t retry_count) : attempted_hosts_(retry_count) {}
 
   bool shouldSelectAnotherHost(const Upstream::Host& candidate_host) override {
     return std::find(attempted_hosts_.begin(), attempted_hosts_.end(), &candidate_host) !=
