+++ source/extensions/retry/priority/previous_priorities/BUILD
@@ -9,10 +9,12 @@ load(
 envoy_package()
 
 envoy_cc_library(
-    name = "other_hosts_predicate_lib",
-    hdrs = ["other_hosts.h"],
+    name = "previous_priorities_lib",
+    srcs = ["previous_priorities.cc"],
+    hdrs = ["previous_priorities.h"],
     deps = [
         "//include/envoy/upstream:retry_interface",
+        "//source/common/upstream:load_balancer_lib",
     ],
 )
 
@@ -21,9 +23,11 @@ envoy_cc_library(
     srcs = ["config.cc"],
     hdrs = ["config.h"],
     deps = [
-        ":other_hosts_predicate_lib",
+        ":previous_priorities_lib",
         "//include/envoy/registry",
         "//include/envoy/upstream:retry_interface",
-        "//source/extensions/retry/host:well_known_names",
+        "//source/common/protobuf",
+        "//source/extensions/retry/priority:well_known_names",
+        "@envoy_api//envoy/config/retry/previous_priorities:previous_priorities_cc",
     ],
 )
