+++ source/extensions/retry/priority/previous_priorities/previous_priorities.h
@@ -9,9 +9,9 @@ namespace Extensions {
 namespace Retry {
 namespace Priority {
 
-class OtherPriorityRetryPriority : public Upstream::RetryPriority {
+class PreviousPrioritiesRetryPriority : public Upstream::RetryPriority {
 public:
-  OtherPriorityRetryPriority(uint32_t update_frequency, uint32_t max_retries)
+  PreviousPrioritiesRetryPriority(uint32_t update_frequency, uint32_t max_retries)
       : update_frequency_(update_frequency) {
     attempted_priorities_.reserve(max_retries);
   }
