+++ test/extensions/retry/priority/previous_priorities/BUILD
@@ -14,9 +14,10 @@ envoy_package()
 envoy_extension_cc_test(
     name = "config_test",
     srcs = ["config_test.cc"],
-    extension_name = "envoy.retry_host_predicates.other_hosts",
+    extension_name = "envoy.retry_priorities.previous_priorities",
     deps = [
-        "//source/extensions/retry/host/other_hosts:config",
+        "//source/extensions/retry/priority:well_known_names",
+        "//source/extensions/retry/priority/previous_priorities:config",
         "//test/mocks/upstream:upstream_mocks",
     ],
 )
